executor/requirements.txt
browser-use>=0.2.0
langchain-openai>=0.3.0
playwright>=1.52.0
fastapi>=0.115.0
uvicorn>=0.34.0
openai>=1.82.0
python-dotenv>=1.1.0
pydantic>=2.11.0
httpx>=0.25.0
